Glock item

+++ src/game/entities/AmmoBox.h
@@ -32,7 +32,7 @@ public:
     static Texture* ms_TextureMinigun;
     // static Sound* ms_PickupSounds[7];
 
-    AmmoBox(GameWorld* world, AmmoType type, double start_x, double start_y, double AmmoCount);
+    AmmoBox(GameWorld* world, AmmoType type, double start_x, double start_y, unsigned int AmmoCount);
 
     // Getting
     [[nodiscard]] AmmoType Type() const { return m_Type; }
